[snapshot import] checkpoint (#27303)
add checkpoints to snapshot import.

1. store the tablet id for any tables so we know which tablet to insert into when a snapshot import restarts.
2. skip the number of documents that already exist in the tablet. this relies on inserting into the database in the same order as the import file, and makes it hard to do parallel transactions, but we don't do those anyway.
3. keep track of whether there are objects without `_id` field, but don't use this info anywhere yet.
4. checkpoints have enough info to construct the "message to confirm" which we also want to display nicely in the dashboard, and allow us to display it after the import has been confirmed.
5. i tested this change by adding random transient errors. see screen recording.
6. fixed the behavior (which also exists in prod) where on transient errors we repeatedly show the message "Imported {table_name}" even if the table name has already been imported.
7. fixed the behavior (which also exists in prod) where we display "Imported {table_name} (1000 documents)" while still displaying a spinner that says "Importing {table_name} (8192 documents)" because this looks really bad in the transient error cases -- we might take a while to make any progress, and in the meantime it shows as if we're continuing to import into a table that has completed.

GitOrigin-RevId: 4c7dabefb1a0cb375419af2a121463056982f360

+++ crates/local_backend/src/import.rs
@@ -85,7 +85,7 @@ enum ImportFormatArg {
 #[derive(Serialize)]
 #[serde(rename_all = "camelCase")]
 struct ImportResponse {
-    num_written: usize,
+    num_written: u64,
 }
 
 fn parse_format_arg(
